Add a class OapOnHeapColumnVector which is subclass of OnHeapColumnVector and implements FiberUsable to avoid adding methods directly to OnHeapColumnVector.
