+++ src/main/java/org/apache/parquet/hadoop/ParquetFiberDataLoader.java
@@ -16,11 +16,11 @@ import org.apache.parquet.hadoop.utils.Collections3;
 import org.apache.parquet.schema.MessageType;
 import org.apache.parquet.schema.Type;
 
-import org.apache.spark.memory.MemoryMode;
 import org.apache.spark.sql.execution.datasources.parquet.ParquetReadSupportWrapper;
 import org.apache.spark.sql.execution.datasources.parquet.VectorizedColumnReader;
 import org.apache.spark.sql.execution.datasources.parquet.VectorizedColumnReaderWrapper;
 import org.apache.spark.sql.execution.vectorized.ColumnVector;
+import org.apache.spark.sql.execution.vectorized.OapOnHeapColumnVector;
 import org.apache.spark.sql.types.DataType;
 import org.apache.spark.sql.types.StructType;
 import org.apache.spark.sql.types.StructType$;
@@ -58,7 +58,7 @@ public class ParquetFiberDataLoader implements Closeable {
     boolean isMissing = isMissingColumn(fileSchema, requestedSchema);
 
     DataType dataType = sparkSchema.fields()[0].dataType();
-    columnVector =ColumnVector.allocate(rowGroupCount, dataType, MemoryMode.ON_HEAP);
+    columnVector = new OapOnHeapColumnVector(rowGroupCount, dataType);
 
     if(isMissing) {
       columnVector.putNulls(0, rowGroupCount);
