+++ src/main/java/org/apache/spark/sql/execution/vectorized/ColumnarBatch.java
@@ -499,7 +499,7 @@ public final class ColumnarBatch {
 
     for (int i = 0; i < schema.fields().length; ++i) {
       StructField field = schema.fields()[i];
-      columns[i] = ColumnVector.allocate(maxRows, field.dataType(), memMode);
+      columns[i] = new OapOnHeapColumnVector(maxRows, field.dataType());
     }
 
     this.row = new Row(this);
